+++ api/src/main/java/oxy/bascenario/api/event/sound/SoundVolumeEvent.java
@@ -5,5 +5,5 @@ import oxy.bascenario.api.effects.Easing;
 import oxy.bascenario.api.event.api.Event;
 
 @Builder(toBuilder = true, builderClassName = "Builder")
-public record SoundVolumeEvent(int id, long duration, float volume, Easing easing) implements Event {
+public record SoundVolumeEvent(int id, int duration, float volume, Easing easing) implements Event {
 }
